Initial commit of a web app to manage my raspberry pi volume
I installed raspotify and found that I needed to ssh into my raspberry
pi to change the volume, as it was a headless setup. Seeing as the
command was pretty simple, I put together a small Go web app to run
the necessary command.

This also includes some configuration for setting things up with
systemd on my raspberry pi.